% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get2CP.R
\name{get2CP}
\alias{get2CP}
\title{Data on two candidate preferred results since 1919}
\usage{
get2CP(include_byelections = F)
}
\arguments{
\item{include_byelections}{Choose to include house by-elections. If .byelections = TRUE,
then a list of elections with by-elections and their divisions is returned.}
}
\value{
a tibble
}
\description{
Data on two candidate preferred results since 1919
}
\examples{
get2CP()
}
