% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommitteeMembers.R
\name{getCommitteeMembers}
\alias{getCommitteeMembers}
\title{Data on committee memberships since 1902}
\usage{
getCommitteeMembers()
}
\value{
a tibble
}
\description{
Data on committee memberships since 1902
}
\examples{
getCommitteeMembers()
}
