% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElections.R
\name{getElections}
\alias{getElections}
\title{Get a list of elections in the Australian Parliament}
\usage{
getElections(chamber = c("all", "house", "senate"), byelections = FALSE)
}
\arguments{
\item{chamber}{select from c("all","house","senate")}

\item{byelections}{choose to include house by elections. if .byelections = TRUE,
then a list of elections with byelections and their divisions will be returned.}
}
\value{
a tibble
}
\description{
Get a list of elections in the Australian Parliament
}
\examples{
getElections(chamber="all")
}
