% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElectorates.R
\name{getElectorates}
\alias{getElectorates}
\title{Data on electorates since 1901}
\usage{
getElectorates()
}
\value{
a tibble
}
\description{
Data on electorates since 1901
}
\examples{
getElectorates()
}
