% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndividuals.R
\name{getIndividuals}
\alias{getIndividuals}
\title{Data on individuals and biographical information since 1901}
\usage{
getIndividuals()
}
\value{
a list
}
\description{
Data on individuals and biographical information since 1901
}
\examples{
getIndividuals()
}
