% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinistries.R
\name{getMinistries}
\alias{getMinistries}
\title{Data on Ministries and Governing Parties since 1901}
\usage{
getMinistries()
}
\value{
a tibble
}
\description{
Data on Ministries and Governing Parties since 1901
}
\examples{
getMinistries()
}
