% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOccupations.R
\name{getOccupation}
\alias{getOccupation}
\title{Data on all listed prior occupations held (both primary and secondary)}
\usage{
getOccupation()
}
\value{
a list
}
\description{
Data on all listed prior occupations held (both primary and secondary)
}
\examples{
getOccupation()
}
