% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOccupations.R
\name{getOccupations}
\alias{getOccupations}
\title{Data on all listed prior occupations held (both primary and secondary)}
\usage{
getOccupations()
}
\value{
a list
}
\description{
Data on all listed prior occupations held (both primary and secondary)
}
\examples{
getOccupations()
}
