% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParliaments.R
\name{getParliaments}
\alias{getParliaments}
\title{Data on Parliamentary terms since 1901}
\usage{
getParliaments()
}
\value{
a tibble
}
\description{
Data on Parliamentary terms since 1901
}
\examples{
getParliaments()
}
