% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParlService.R
\name{getParlService}
\alias{getParlService}
\title{data on parliamentary tenures of MPs and senators since 1901}
\usage{
getParlService(chamber = c("all", "house", "senate"))
}
\arguments{
\item{chamber}{choose to include MPs, senators or both.}
}
\value{
a tibble
}
\description{
up to date as of 13th December 2022
}
\examples{
getParlService(chamber="house")
}
