% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPartyService.R
\name{getPartyService}
\alias{getPartyService}
\title{data on party service of MPs and senators since 1901}
\usage{
getPartyService()
}
\value{
a tibble
}
\description{
up to date as of 9th March 2024
}
\examples{
getPartyService()
}
