% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQualifications.R
\name{getQualifications}
\alias{getQualifications}
\title{Data on all listed Qualifications}
\usage{
getQualifications()
}
\value{
a list
}
\description{
Data on all listed Qualifications
}
\examples{
getQualifications()
}
