% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShadowMinisters.R
\name{getShadowMinisters}
\alias{getShadowMinisters}
\title{Data on Shadow Ministers since 1901}
\usage{
getShadowMinisters()
}
\value{
a tibble
}
\description{
Data on Shadow Ministers since 1901
}
\examples{
getShadowMinisters()
}
