% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSittingDates.R
\name{getSittingDates}
\alias{getSittingDates}
\title{data on sitting dates in the federal parliament}
\usage{
getSittingDates()
}
\value{
a tibble
}
\description{
data on sitting dates in the federal parliament
}
\examples{
getSittingDates()
}
