% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElections.R
\name{getElections}
\alias{getElections}
\title{Data on elections since 1901}
\usage{
getElections(chamber = c("all", "house", "senate"), byelections = FALSE)
}
\arguments{
\item{chamber}{Select from c("all","house","senate")}

\item{byelections}{Choose to include house by-elections. If .byelections = TRUE,
then a list of elections with by-elections and their divisions is returned.}
}
\value{
a tibble
}
\description{
Data on elections since 1901
}
\examples{
getElections(chamber="all")
}
