% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParliaments.R
\name{getParliaments}
\alias{getParliaments}
\title{Get list of Parliaments, their start date, dissolution date, end date and election date}
\usage{
getParliaments()
}
\value{
a tibble
}
\description{
Get list of Parliaments, their start date, dissolution date, end date and election date
}
\examples{
getParliaments()
}
